# docs/documentation-implementation/api-reference/openapi.yaml
openapi: 3.1.0
info:
  title: SkillSwap API
  description: |
    API REST de SkillSwap - Plateforme d'échange de compétences entre utilisateurs.

    ## Authentification
    L'API utilise des tokens JWT stockés dans des cookies HTTP-only sécurisés.
    - `accessToken` : valide 15 minutes, utilisé pour authentifier les requêtes
    - `refreshToken` : valide 7 jours, utilisé pour renouveler l'accessToken

    ## Base URL
    - **Docker (via Nginx)** : `http://localhost:8888/api/v1`
    - **Backend direct** : `http://localhost:3000/api/v1`
    - **Production** : `https://api.skillswap.fr/api/v1`

    ## Codes d'erreur
    | Code | Description |
    |------|-------------|
    | 400 | Données de requête invalides |
    | 401 | Non authentifié ou token expiré |
    | 403 | Accès refusé (permissions insuffisantes) |
    | 404 | Ressource non trouvée |
    | 409 | Conflit (ressource déjà existante) |
    | 500 | Erreur serveur interne |
  version: 1.0.0
  contact:
    name: Équipe SkillSwap
    url: https://github.com/O-clock-Dublin/projet-skillswap
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8888/api/v1
    description: Développement local (Docker via Nginx)
  - url: http://localhost:3000/api/v1
    description: Backend direct (sans Nginx)
  - url: https://api.skillswap.fr/api/v1
    description: Production
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rafraîchir le token
      description: Génère un nouveau accessToken à partir du refreshToken.
      operationId: refreshToken
      security: []  # Uses refreshToken cookie, not accessToken
      responses:
        '200':
          description: Token rafraîchi
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: Refresh token invalide ou expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /profiles/{id}:
    get:
      tags: [Profiles]
      summary: Détail d'un profil
      description: Retourne les informations complètes d'un utilisateur.
      operationId: getProfile
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Profil trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          description: Utilisateur non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Profiles]
      summary: Modifier son profil
      description: Met à jour partiellement les informations du profil (propriétaire uniquement).
      operationId: updateProfile
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profil mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '403':
          description: Non autorisé (pas le propriétaire)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /profiles:
    delete:
      tags: [Profiles]
      summary: Supprimer son compte
      description: Supprime définitivement le compte de l'utilisateur connecté.
      operationId: deleteAccount
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Compte supprimé
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /profiles/password:
    patch:
      tags: [Profiles]
      summary: Changer son mot de passe
      description: |
        Permet à l'utilisateur connecté de modifier son mot de passe.

        **Validation :**
        - L'ancien mot de passe doit être correct
        - Le nouveau mot de passe doit contenir au moins 8 caractères
        - La confirmation doit correspondre au nouveau mot de passe
      operationId: changePassword
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            example:
              currentPassword: "AncienMotDePasse123!"
              newPassword: "NouveauMotDePasse456!"
              confirmation: "NouveauMotDePasse456!"
      responses:
        '200':
          description: Mot de passe modifié avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Mot de passe modifié avec succès"
        '400':
          description: Données invalides (confirmation ne correspond pas)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: L'ancien mot de passe est incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================
  # SKILLS ENDPOINTS
  # ============================================
  /profiles/skills:
    get:
      tags: [Skills]
      summary: Liste des compétences
      description: Retourne toutes les compétences disponibles dans le système.
      operationId: getAllSkills
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Liste des compétences
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Skill'

    post:
      tags: [Skills]
      summary: Ajouter une compétence
      description: Ajoute une compétence au profil de l'utilisateur connecté.
      operationId: addSkill
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skillId]
              properties:
                skillId:
                  type: integer
                  example: 1
      responses:
        '201':
          description: Compétence ajoutée
        '400':
          description: Compétence déjà présente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /profiles/availabilities:
    get:
      tags: [Availabilities]
      summary: Liste des disponibilités
      description: Retourne tous les créneaux de disponibilité possibles.
      operationId: getAllAvailabilities
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Liste des disponibilités
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Availability'

    post:
      tags: [Availabilities]
      summary: Ajouter une disponibilité
      description: Ajoute un créneau de disponibilité au profil.
      operationId: addAvailability
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityRequest'
      responses:
        '201':
          description: Disponibilité ajoutée
...
  /profiles/avatar:
    patch:
      tags: [Profiles]
      summary: Mettre à jour l'avatar
      description: |
        Upload d'une nouvelle image de profil.

        **Formats acceptés** : JPEG, JPG, PNG
        **Taille max** : 5 MB
      operationId: updateAvatar
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar:
                  type: string
                  format: binary
                  description: Fichier image (JPEG, JPG ou PNG, max 5MB)
      responses:
        '200':
          description: Avatar mis à jour
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      avatarUrl:
                        type: string
                        format: uri
                        example: "/avatars/avatar-1-1234567890.jpg"
        '400':
          description: Format de fichier invalide ou fichier trop volumineux
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Fichier trop volumineux (> 5MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Profiles]
      summary: Supprimer l'avatar
      description: Supprime l'image de profil et remet l'avatar par défaut.
      operationId: deleteAvatar
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Avatar supprimé
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Avatar supprimé avec succès"

  # ============================================
  # PUBLIC PROFILE ENDPOINT (SEO)
  # ============================================
  /profiles/public/{id}:
    get:
      tags: [Profiles]
      summary: Profil public (SEO)
      description: |
        Retourne les informations publiques d'un profil pour le SEO et les réseaux sociaux.

        **Aucune authentification requise.**

        Utilisé pour :
        - Génération des méta-tags OpenGraph
        - Aperçu sur les réseaux sociaux
        - Indexation par les moteurs de recherche
      operationId: getPublicProfile
      security: []  # Public endpoint - no authentication required
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Profil public
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicProfile'
        '404':
          description: Utilisateur non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

# ============================================
# COMPONENTS
# ============================================
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: JWT token stocké dans un cookie HTTP-only
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: ID de l'utilisateur
...
  schemas:
...
    UserProfile:
      type: object
      description: Profil complet d'un utilisateur
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          description: Email (visible uniquement pour son propre profil)
          example: "marie@example.com"
        firstname:
          type: string
          example: "Marie"
        lastname:
          type: string
          example: "Dupont"
        address:
          type: ["string", "null"]
          example: "15 rue de la Paix"
        postalCode:
          type: ["integer", "null"]
          example: 75001
        city:
          type: ["string", "null"]
          example: "Paris"
        age:
          type: ["integer", "null"]
          example: 28
        avatarUrl:
          type: ["string", "null"]
          format: uri
          example: "https://example.com/avatar.jpg"
        description:
          type: ["string", "null"]
          example: "Passionné de développement web"
        skills:
          type: array
          description: Compétences de l'utilisateur
          items:
            $ref: '#/components/schemas/UserSkill'
        interests:
          type: array
          description: Compétences recherchées par l'utilisateur
          items:
            $ref: '#/components/schemas/UserSkill'
        availabilities:
          type: array
          description: Créneaux de disponibilité
          items:
            $ref: '#/components/schemas/UserAvailability'
        evaluationsReceived:
          type: array
          description: Évaluations reçues par l'utilisateur
          items:
            $ref: '#/components/schemas/RatingReceived'
        followedUsers:
          type: array
          description: Utilisateurs suivis
          items:
            $ref: '#/components/schemas/Follow'
        followerUsers:
          type: array
          description: Utilisateurs qui suivent ce profil
          items:
            $ref: '#/components/schemas/Follow'
        isRated:
          type: boolean
          description: Indique si l'utilisateur connecté a déjà noté ce profil (uniquement sur profil d'autrui)
          example: false
        isFollowing:
          type: boolean
          description: Indique si l'utilisateur connecté suit ce profil (uniquement sur profil d'autrui)
          example: true
        isFollowedBy:
          type: boolean
          description: Indique si ce profil suit l'utilisateur connecté (uniquement sur profil d'autrui)
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-10T08:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T14:30:00Z"
    UserSkill:
      type: object
      description: Compétence associée à un utilisateur
      properties:
        userId:
          type: integer
          example: 1
        skillId:
          type: integer
          example: 5
        skill:
          $ref: '#/components/schemas/Skill'
        createdAt:
          type: string
          format: date-time
    UserAvailability:
      type: object
      description: Disponibilité associée à un utilisateur
      properties:
        userId:
          type: integer
          example: 1
        availableId:
          type: integer
          example: 3
        available:
          $ref: '#/components/schemas/Availability'
        createdAt:
          type: string
          format: date-time
    RatingReceived:
      type: object
      description: Évaluation reçue avec informations sur l'évaluateur
      properties:
        id:
          type: integer
          example: 1
        score:
          type: integer
          minimum: 0
          maximum: 5
          example: 4
        comments:
          type: ["string", "null"]
          example: "Excellent échange !"
        evaluatorId:
          type: integer
          example: 2
        evaluatedId:
          type: integer
          example: 1
        evaluator:
          $ref: '#/components/schemas/UserSummary'
        createdAt:
          type: string
          format: date-time
    Follow:
      type: object
      description: Relation de suivi entre utilisateurs
      properties:
        id:
          type: integer
          example: 1
        followerId:
          type: integer
          example: 1
        followedId:
          type: integer
          example: 2
        createdAt:
          type: string
          format: date-time
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        firstname:
          type: string
        lastname:
          type: string
        avatarUrl:
          type: ["string", "null"]
    UpdateProfileRequest:
      type: object
      description: Données pour mettre à jour le profil utilisateur
      properties:
        lastname:
          type: string
          description: Nom de famille
          pattern: "^[A-Za-zÀ-ÿ\\s\\-']+$"
          example: "Dupont"
        firstname:
          type: string
          description: Prénom
          pattern: "^[A-Za-zÀ-ÿ\\s\\-']+$"
          example: "Marie"
        email:
          type: string
          format: email
          description: Adresse email
          example: "marie.dupont@example.com"
        address:
          type: string
          description: Adresse postale
          minLength: 5
          maxLength: 100
          example: "15 rue de la Paix"
        postalCode:
          type: integer
          description: Code postal (5 chiffres)
          minimum: 1000
          maximum: 99999
          example: 75001
        city:
          type: string
          description: Ville
          example: "Paris"
        age:
          type: integer
          description: Âge de l'utilisateur
          minimum: 1
          example: 28
        avatarUrl:
          type: string
          description: URL de l'avatar
          maxLength: 200
          example: "https://example.com/avatar.jpg"
        description:
          type: string
          description: Description du profil
          minLength: 10
          maxLength: 500
          example: "Passionné de développement web, j'aime partager mes connaissances."
    ChangePasswordRequest:
      type: object
      description: Données pour changer le mot de passe
      required: [currentPassword, newPassword, confirmation]
      properties:
        currentPassword:
          type: string
          description: Mot de passe actuel
          minLength: 8
          example: "AncienMotDePasse123!"
        newPassword:
          type: string
          description: Nouveau mot de passe (minimum 8 caractères)
          minLength: 8
          example: "NouveauMotDePasse456!"
        confirmation:
          type: string
          description: Confirmation du nouveau mot de passe
          example: "NouveauMotDePasse456!"

    # ---- Skill Schemas ----
    Skill:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "React"
        categoryId:
          type: integer
          example: 1
        category:
          $ref: '#/components/schemas/Category'
    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Développement Web"
        slug:
          type: string
          example: "developpement-web"

    # ---- Availability Schemas ----
    Availability:
      type: object
      description: Créneau de disponibilité d'un utilisateur
      properties:
        id:
          type: integer
          example: 1
        day:
          type: string
          description: Jour de la semaine (en français)
          enum: [Lundi, Mardi, Mercredi, Jeudi, Vendredi, Samedi, Dimanche]
          example: Lundi
        timeSlot:
          type: string
          description: Créneau horaire
          enum: [Morning, Afternoon]
          example: Morning
    AvailabilityRequest:
      type: object
      description: Données pour ajouter une disponibilité
      required: [day, timeSlot]
      properties:
        day:
          type: string
          description: Jour de la semaine (en français)
          enum: [Lundi, Mardi, Mercredi, Jeudi, Vendredi, Samedi, Dimanche]
          example: Mardi
        timeSlot:
          type: string
          description: Créneau horaire (Morning = Matin, Afternoon = Après-midi)
          enum: [Morning, Afternoon]
          example: Afternoon

    # ---- Rating Schemas ----
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Email déjà utilisé"
        code:
          type: string
          example: "CONFLICT"

    # ---- Search Schemas ----
...
    PublicProfile:
      type: object
      description: Profil public pour le SEO (sans données sensibles)
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            id:
              type: integer
              example: 1
            firstname:
              type: string
              example: "Marie"
            lastname:
              type: string
              example: "Dupont"
            city:
              type: ["string", "null"]
              example: "Paris"
            avatarUrl:
              type: ["string", "null"]
              format: uri
              example: "/avatars/avatar-1-123456.jpg"
            description:
              type: ["string", "null"]
              example: "Développeuse passionnée"
            skills:
              type: array
              items:
                $ref: '#/components/schemas/UserSkill'
            evaluationsReceived:
              type: array
              items:
                $ref: '#/components/schemas/RatingReceived'
            createdAt:
              type: string
              format: date-time